SERVER-100901 Relax user digest invariant to tassert in ShardingTaskExecutor (#32452)
The invariant on a matching `uid` in the `lsid` of the `opCtx` and
command request in `sharding_task_executor.cpp` has led to several AFs
involving server crashes. While this is a sign of unintended behavior,
it is not so serious that it should be crashing the server. We should
log this as a programmer error instead.