SERVER-95435 Move OpDebug class to own file (#32236)
This change moves the OpDebug class to its own file in an attempt to
split up curop.cpp and curop.h into more manageable files. Helper
functions such as `appendCommentField`, `buildTruncatedObject`, and
`appendObjectTruncatingAsNecessary` that used to be in the anonymous
namespace in curop.cpp are now in the curop_util namespace and defined
in op_debug.h to be reused in op_debug.cpp and curop.cpp.