SERVER-99169: Validate references to "score" meta field (#32231)
Enables validation of `{$meta: "score"}`. That means if you try to
access "score" metadata in a pipeline that doesn't generate it, the
query will error at parse time from the `DepsTracker`.

This also required a minor change to $rankFusion desugaring to avoid
putting `{$meta: "score"}` for scoreDetails in the output pipeline
unless the pipeline does actually generate a score. That doesn't change
the scoreDetails output though, since in its current state, the "score"
would evaluate to missing.

### Testing
Most of the test coverage comes from modifying and expanding
`meta_dependency_validation.js`: I added the usage of `$vectorSearch`
and `$score` stages for comprehensive testing of "score".