SERVER-100727 Check for consumed _cmdRequest in FindCmd::explain (#32270)
This is addressing the coverity UBSAN failure in
https://jira.mongodb.org/browse/BF-36521

I wanted to open a draft request first to discuss if this is the
appropriate resolution for this, or if we should expect that explain()
will not be called in such a state, and if so we should instead simply
be asserting against having a null value. Also, I noted that there don't
seem to be unit tests covering find_cmd.cpp, are they in another
location I missed?