SERVER-99246 Implement ConnectionPoolStats in gRPC AsyncClientFactory (#32157)
This commit implements the `appendStats` method in the gRPC
`AsyncClientFactory` with the `ConnectionPoolStats` tracking information
about gRPC streams within the factory. The only exception to this is the
`ConnectionPoolStats::totalAvailable` stat which tracks the total number
of gRPC channels that are open. This is because the available stat has a
connection pooling specific definition which does not fit nicely with
gRPC, so I ended up implementing what would match its intended
definition the closest.