SERVER-93786: Print explain diagnostics during tassert/invariant/segfault failure during query execution (#31457)
After query planning is complete, a 'ScopedDebugInfo' captures a
reference to the query executor, allowing us to log the winning plan and
execution stats of the executor if we encounter a
tassert/invariant/segfault.

This commit includes the following changes:
- A new diagnostic printer, 'ExplainDiagnosticPrinter', is introduced. 
- Query commands build a 'ScopedDebugInfo' for the
'ExplainDiagnosticPrinter' once they have an executor
- 'PlanExplainerPipeline' implements a basic form of
'getWinningPlanStats()' which reports the winning plan stats from its
cursor, if one exists.
- Some invariants in 'PlanExplainer's are replaced with tasserts