SERVER-98646 Fix insufficient distinct strings values in histogram benchmark (#30506)
The data interval for strings is too narrow for `ndv` as it represents
string lengths. The small `ndv` derived makes the benchmark results show
negligible impact from the number of buckets. Therefore, that interval
is unsuitable for calculating `ndv`.

This commit sets `ndv` to `numberOfQueries` to ensure we have enough
distinct values for histograms to partition buckets.