SERVER-100394: Validate "score" $meta dependencies in mongot queries (#32436)
Adds validation of "score" $meta field dependencies on the path when
preparing for mongot queries. In turn, this re-allowed access to "score"
meta field for sharded {$search} and {$vectorSearch} queries.

This PR intentionally maintains the current behavior for
"geoNearDistance" and "geoNearPoint" $meta field dependencies, which are
_not_ validated (and therefore are misleading allowed) for mongot
queries. Even though that change would now come "for free", we will
postpone that change to isolate any backwards-breaking changes. This
doesn't apply to "score" since it is a new feature-flag-guarded field.