SERVER-101009 Add support for RHEL 9 PPC (#32538)
Add support for RHEL 9 PPC to master branch

In RHEL8 and RHEL9 the debug symbols for libgcc aren't stripped and are
instead split, which still leaves behind debug symbols in the libgcc
shared object file. These debug symbols are created with gdwarf32, so
they're limited to a 32 bit address space. Even if the mongodb binaries
are compiled with gdwarf64, there's a chance that the gdwarf32 libgcc
debug symbols will be placed after the gdwarf64 debug symbols. This
started happening in the RHEL9 ppc64le build.

The workaround for this is stripping the debug symbols from libgcc and
statically linking the libgcc from the toolchain into the mongodb
binaries. The longer term solution for this is to containerize the
non-remote-execution build and strip the debug symbols inside the
container, or patch the compilers to properly order gdwarf32 symbols
before gdwarf64 symbols. See https://reviews.llvm.org/D96144 for the
possible compiler patch to fix the issue

https://jira.mongodb.org/browse/SERVER-101099 tracks the longer term
solution of containerizing the system libraries in non-remote-execution
builds