SERVER-100001: Created script to repro evergreen failures locally (#32341)
I found that editing the test corpus based on the evergreen failures was
quite tedious and involved a really convoluted workflow.
This PR introduces 2 scripts:
- setup_repro_env.py
- overwrite_results.py
See the comment below for the draft doc on how to use these scripts.

__Goal__: Minimize friction between seeing the test failures on
evergreen and having them locally.

The `setup_repro_env.py` script takes a set of evergreen task_ids as
inputs and handles sparse-checking out all the appropriate failing test,
results and coll files. It also outputs a formatted string that can be
used in the invocation of `mongotest`. They developer could also use the
resmoke invocation given by evergreen (pending
[SERVER-100813](https://jira.mongodb.org/browse/SERVER-100813)).

From this point on, the developer is able to choose a few paths.
1. they can use the `overwrite_results.py` script to overwrite the
.results files with the .actual files and run a `git diff` / `git add
-p` inside of each of the test corpus repos to choose which diffs/files
to accept
2. Manually run `git diff --no-index` on the .results and .actual files
to see the diffs.
3. Re-record the results entirely. We could consider adding a resmoke
suite that allows for this.
[SERVER-100814](https://jira.mongodb.org/browse/SERVER-100814) tracks
this work.