SERVER-96197 Use nss instead of coll names for resolved namespace map keys (#32165)
This pr changes ExpressionContext's resolvedNamespaces member variable
to be keyed on `NamespaceString`'s instead of `std::string`'s
representing collection/view names. Although collection/view names are
guaranteed to be unique within a single database, this restriction does
not apply across databases. This meant that when you had a cross
database operation such as a pipeline with a `$lookup` and a
cross-database `$out` or `$merge`, `resolvedNamespaces` could be
populated incorrectly if the collections/views in the different
databases had the same name. Since `NamespaceString`'s also include the
database name, two collections/views with the same name but in different
databases will now have distinct keys in the map.

There is also a new `allowGenericForeignDbLookup` flag added to
`ExpressionContext`. This flag is required for "heavy" parsing
(`createFromBson`) the special non-MQL `$lookup` syntax used by streams
and should not be enabled by any other code.

There is also a new `LiteParserOptions` struct that is used an additonal
argument to lite parsing. Currently this contains a new
`allowGenericForeignDbLookup` flag so that we can properly determine if
a $lookup stage is allowed to use `from: {db: ..., coll: ...}` syntax
and throw errors when it isn't allowed instead of deferring to the
"heavy" parsing in `createFromBson`.