SERVER-97459: introduce type declarations for global JS objects (#32684)
This provides hover-over hints in the IDE where there was no
traceability before (everything was just `any` with no descriptions).

I framed out methods on `assert` and others to prove the plumbing
worked, and added rich details in some spots to demonstrate more ideal
cases of what power can be leveraged with this. This pattern can be
referenced and ticketed out to cover other globals and enrich type help
with embedded documentation.