SERVER-100242 Add Parsing section to QO Architecture Guide (#32458)
A couple organization questions I ran into while writing: 

1. CanonicalQuery is "owned" by logical models, but we need to explain
what it is here. I've written it up here, but maybe we could add a link
to it in the logical models section instead?
2. I think we're better off with de-sugaring and serializing in logical
models, under `MatchExpression`. Thoughts?
3. I found an in-depth Stable API README
[here](https://github.com/10gen/mongo/blob/master/src/mongo/db/STABLE_API_README.md).
I don't see any value in re-writing this, so I added a link to it in our
Table of Contents. What do you think of this?