SERVER-101233 compile resharding_coordinator.cpp in parts using ifdefs (#32758)
This PR proposes that we can keep the resharding coordinator's
implementation in a single file, but compile it in parts by using
`#ifdef` directives. Each part of the main `.inl` file is compiled by a
stub `.cpp` file that does nothing but
set a `#define` selecting a part number to build, and then `#include`
the `.inl` file.

This is a replacement for #32707, which split the
`resharding_coordinator.cpp` file into multiple .cpp file parts.
That was understandably considered too disruptive a change, because it
becomes difficult to see and modify relationships among the parts of the
class implementation.