SERVER-101223 Add linter for missing scons test stubs (#32696)
Add a linter to "bazel run lint" that verifies that all Bazel unittest
targets have a SCons unit test stub to trigger execution in CI. This
also adds stubs for the handful of existing tests that don't have a
SCons stub.

This is temporary while we're still invoking unittests through the
SCons/resmoke interface. After migrating to use "bazel test" in CI this
will no longer be needed.