SERVER-99569 Validate partitionBy and sortBy in $fill (#32252)
In the scenario where `$setWindowFields` is optimized out of `$fill`
during the rewrite, the `sortBy` and `partitionBy` parameters would not
be parsed and validated. This would allow a user the ability to pass in
bogus expressions. In order to solve this we need to parse these
expressions in the `$fill` stage if we recognise that `$setWindowFields`
will not be used.