SERVER-100971 Fix dangling pointer to OperationContext in PipelineOptimizationTest (#32766)
`assertPipelineOptimizesTo` creates an `OperationContext`, and then a
`Pipeline` with a reference to it. At the end of the method, it returns
that `Pipeline`, but the locally-scoped `OperationContext` is destroyed
there. This leaves `Pipeline` with a dangling `OperationContext`
pointer, which would cause a use-after-free if it later tries to use
that `OperationContext`.

(This was found by ASAN on `db_pipeline_test`, while prototyping feature
flag checking against information associated to the `OperationContext`.)