SERVER-101094 Add kCachePressure rollover case for _generateMeasurementsWithRolloverReason (#32790)
- Made `kStorageCacheSizeBytes` **not** a `static constexpr` in the
`bucket_catalog_tests.cpp` and `timeseries_write_ops_internal.cpp`, but
a private variable (`_storageCacheSizeBytes`) that is initialized during
the `setUp()` of the test fixtures assigned to
`representativeStorageCacheSizeBytes`
- Created` static constexpr` variables
`representativeStorageCacheSizeBytes` and
`storageCacheSizeBytesForCachePressure`
- Created unit tests that exercise `kCachePressure` rollover reasons for
`stageInsertBatchIntoEligibleBucket` and `stageInsertBatch`.

Issues:
- We can't add an invariant in `_testStageInsertBatchIntoEligibleBucket`
and `_testStageInsertBatch` on having the correct
`_storageCacheSizeBytes` set because we are blind to what is in the
measurements vector (what type of rollover reasons there can be). [Is
this
comment](https://github.com/10gen/mongo/pull/32790/files#diff-3005f774ab340e872dd1ef7a4699095988f6ae5bf7476326331a83ac6101fa37R369)
(same comment before both `_testStageInsertBatch` and
`_testStageInsertBatchIntoEligibleBucket`) sufficient to address that
these are responsibility of the user doing the unit tests? Are there any
other ways we can enforce this?