SERVER-98461 Return WCE for non-shard key findAndModify (#32713)
Make sure that the `writeConcernError` field is returned properly from
cluster non-shard key `findAndModify` commands.
Non-shard key write commands are executed in an internal transaction,
which first looks up the document on all candidate shards and then
issues the write command to the target shard the document was found on.
If the write succeeds, the internal transaction is committed.
This functionality is encapsulated internally in the
`runTwoPhaseWriteProtocol`
[function](https://github.com/10gen/mongo/blob/ac3a9ec25fee78b30cd4af625b3ebd8f615ef80e/src/mongo/s/write_ops/write_without_shard_key_util.cpp#L281-L398).
At the end of this function the [effective status of the internal
transaction's commit result is
fetched](https://github.com/10gen/mongo/blob/ac3a9ec25fee78b30cd4af625b3ebd8f615ef80e/src/mongo/s/write_ops/write_without_shard_key_util.cpp#L388-L394).
This call [can clobber any writeConcernError that was also present in
the transaction's commit
result](https://github.com/10gen/mongo/blob/ac3a9ec25fee78b30cd4af625b3ebd8f615ef80e/src/mongo/db/transaction/transaction_api.h#L93-L102),
in case an additional error was registered in the transaction. Only the
transaction's final status is then returned as the result of
`runTwoPhaseWriteProtocol`. So if there is an error registered in the
transaction and a writeConcernError happened, the writeConcernError is
discarded.

This PR changes the return value of `runTwoPhaseWriteProtocol` so that
does not clobber a potential writeConcernError. If a write operation
runs into an error and also a writeConcernError, the error is now
wrapped into an `ErrorWithWriteConcernError`, which contains both
components.
The PR also adjusts two places utilized by write operations (BulkWrite
and WriteError) to serialize errors to BSON. Here we need to perform an
extra check if the error is an `ErrorWithWriteConcernError`. This is
needed so that we return the original error code and message in addition
to the writeConcernError. Not doing that would lead to a change in error
codes if a writeConcernError occurs.

---------

Co-authored-by: jannaerin <golden.janna@gmail.com>
Co-authored-by: Abdul Qadeer <abdul.qadeer@mongodb.com>