SERVER-99153: Account for stages that drop metadata when validating "score" and "scoreDetails" $meta references (#32804)
This PR expands on $meta field validation by accounting for stages that
lose per-document metadata (like $group) when validating "score" and
"scoreDetails". For example, a pipeline like `[{$search}, {$group},
{$project "scoreDetails"}]` will now throw a uassert, alerting the user
that "scoreDetails" is not available when requested (since the
"scoreDetails" is lost at the $group stage).

As a side effect, we're able to correctly validate references to
"scoreDetails" after $rankFusion pipelines.

**This PR does not change the behavior of validation for any other
metadata field.** That work is tracked by WRITING-29625, separately
since those changes may be backwards-breaking.