SERVER-100299 modified pipeline optimization for sort-skip-limit on time-series collections to return correct number of documents (#32733)
During optimization of sort-skip-limit queries on time-series
collections, `$limit` gets modified to equal the sum of values from
`$limit` and `$sort`. This updated value is absorbed by `$sort`. In
previous implementation, the updated `$limit` was pushed at the end of
the pipeline resulting in incorrect number of documents being return as
`$skip` came before `$limit`. To fix this issue, the optimization logic
was modified to push the updated `$limit` immediately after `$sort`
resulting in `$limit` being followed by `$skip` which returns the
correct number of documents.