SERVER-101277 Update _generateMeasurementsWithRolloverReason to accept more optional args (#32750)
- Added more optional inputs for
`_generateMeasurementsWithRolloverReason` including:
- `numMeasurements` (for `RolloverReason` `kSchemaChange`,
`kTimeForward`, `kTimeBackward`, and `kNone`); other rollover reasons
depend on a specific `numMeasurement` to simulate the `RolloverReason`
- `idxOfDiffMeasurement` which specifies at what index in the batch we
begin to have a measurement that can lead to the rollover reasons
`kSchemaChange`, `kTimeForward`, `kTimeBackward`
- `metaValue` so we can do more testing within a single unit test (so
that we can make vectors with meta values that won't be written to the
same buckets)
- `timeValue` so we can do more testing of mixing different
`RolloverReason`s in the same `BatchedInsertContext` (ensures that we
know the ordering of the different vectors that are being "mixed
together" in the same `BatchedInsertContext` with different
`RolloverReason`s.)
- Update current testing to correctly interact with the new
`_generateMeasurementsWithRolloverReason`
- Added testing for having multiple `RolloverReason`s in the same
`BatchedInsertContext` for `stageInsertBatch` and
`insertBatchIntoEligibleBucket`
- Added TODOs for
[SERVER-101094](https://jira.mongodb.org/browse/SERVER-101094)
    
Note: I'll talk about `kCachePressure` but we aren't currently testing
this right now. However, this will be addressed next up in
[SERVER-101094](https://jira.mongodb.org/browse/SERVER-101094). We will
still use the same vector generated by
`_generateMeasurementsWithRolloverReason` for the `kSize` case and the
`kCachePressure` case because we can use this same vector while lowering
the `kStorageCacheSizeBytes` in the testing files (example in
timeseries_write_ops_internal_test.cpp
[here](https://github.com/10gen/mongo/pull/32750/files#diff-3005f774ab340e872dd1ef7a4699095988f6ae5bf7476326331a83ac6101fa37R51))
to trigger the `kCachePressure` case (supposedly).

Some "issues" (I put these in quotes because I'm not sure how much these
are concern/if these are worth investing to fix) with
`_generateMeasurementsWithRolloverReason`:
1. Fixed number of measurements for `RolloverReason`s `kSize`,
`kCachePressure` (both 125) and `kCount ` (`2 *
gTimeseriesBucketMaxCount`)
2. Fixed `idxOfDiffMeasurement` for the`RolloverReason`s `kSize` and
`kCachePressure`. We always have the case that the 125 (the last element
in the vector) will trigger the `kSize`/`kCachePressure` case.
3. I currently do not love the way I am declaring the mixed rollover
reason vectors (see
[here](https://github.com/10gen/mongo/pull/32750/files#diff-3005f774ab340e872dd1ef7a4699095988f6ae5bf7476326331a83ac6101fa37R670)),
specifically just thinking it looks a little clunky.

Potential ways of solving these "issues":
- For the `numMeasurements` arg, we take in a multiplier value for
`kSize`, `kCachePressure`, and `kCount`. So in the current code, the
vector generated by `_generateMeasurementsWithRolloverReason` with
`kCount`'s size will be `numMeasurements * gTimeseriesBucketMaxCount`
and the vector generated by `_generateMeasurementsWithRolloverReason`
size will be `numMeasurements * 125`. This would also solve issue 2
above because if we could strategically make the multiplier have it be
so that we don't always have the last measurement for
`kSize`/`kCachePressure` trigger the rollover.
- Add an equation that will enable us to create dynamically allocate our
large vector that would trigger the `kSize` and `kCachePressure`
conditions given an input number of measurements that will also enable
us to use `idxOfDiffMeasurement` and enable more coverage in unit
testing. I don't love this idea just because I believe it will make
these unit tests less digestible.
- I'm open to any suggestions for the last bullet, honestly not 100%
sure with how to make this look better..

---------

Co-authored-by: Damian Wasilewicz <33820523+DamianWasilewicz@users.noreply.github.com>