SERVER-93816 Allow core property-based tests to minimize documents (#32471)
To run a property-based test, we now need a collection model and an agg
model. The collection model includes indexes, documents, and whether the
collection is time-series or not. Documents are now modeled in
fast-check.

Minimization now takes longer since the indexes, documents, and queries
all have to be minimized. But it's ok to be slow in the failure case.

This also will help enable
[SERVER-94730](https://jira.mongodb.org/browse/SERVER-94730), since
we'll have the collection details all in one place, rather than having
hard-coded documents that the PBT runner doesn't know about