SERVER-101248 Error message is missing when exceptions are thrown in jstests (#32726)
We are removing the `AssertionError` class and instead add `extraAttr`
to Error because it already has other attributes added to it (and
otherwise we'd need to add them to AssertionError too).

Additionally, when logFormat is json, we add the `errmsg` from command
result (if it exists) to the error message of the assertions because
some tests assert on specific message.