SERVER-100943: Fix abort transaction behavior with WriteConcernExceptions (#32749)
Currently if we have a transaction that has been committed we do not
wait for write concern to make sure that commitTransaction has actually
satisfied writeConcern. This can lead to us saying a transaction was
committed when we try to run abortTransaction when actually the
commitTransaction did not satisfy write concern and might end up being
rolled back.

This change forces us to wait for the commit transaction to be
replicated, matching the writeConcern of the abort transaction command