SERVER-101623 move _makeStageImpl into sort_stage_sort_impl.cpp (#33093)
A simplification of prior SERVER-101623 commit #32980.

We can just move the `_makeStageImpl`. Then there's no need to use
`extern template`.
All of the `SortImpl` usage is contained to the
`sort_stage_sort_impl.cpp` file, which no longer needs a header.
We can keep all references to `SortImpl` contained to that one file.