SERVER-100260 Support using unencrypted FLE1 schemas with QE in lookup (#32862)
This change enables mixing FLE2 encryption schemas with unencrypted
collections. In general, query analysis must always receive a schema for
every collection referenced in the command, as this guarantees it has
all required information to properly analyze the query.

In $lookup scenarios involving both CSFLE schemas and unencrypted
collections, the unencrypted collections provide an entry in
`csfleEncryptionSchemas` with an empty json schema. However, in
scenarios involving both FLE2 schemas and unencrypted collections, the
unencrypted collections must provide their empty schema under
`csfleEncryptionSchemas`. This is because `encryptionInformation` , is
reserved exclusively for FLE2 encryption schemas, and does not support
empty schemas.

This means that such cases, a query would contain both
`encryptionInformation` and `csfleEncryptionSchemas` containing the
unencrypted schemas. It must be emphasized that under no circumstances
is mixing of both CSFLE and FLE2 in a single query supported. In this
change, query analysis guarantees that in the event that a query
contains both `encryptionInformation` and `csfleEncryptionSchemas`, all
the schemas in `csfleEncryptionSchemas` are unencrypted.

In contrast to query analysis, when a FLE2 query reaches the server, it
only contains the FLE2 encryption schemas as `encryptionInformation`.
When a `QueryRewriter` is instantiated, it must be provided with the ESC
collection's namespace corresponding to the pipeline's involved
collection. In the case of an unencrypted collection, there is no such
ESC collection's namespace to be found in `encryptionInformation`. Since
a `QueryRewriter` must always be instantiated to rewrite each nested
sub-pipeline in an aggregation, we provide an empty namespace to
indicate that the `QueryRewriter` corresponds to a pipeline involving an
unencrypted collection.

In the unlikely event that the `QueryRewriter` for a pipeline involving
an unencrypted collection contains encryption payloads that need to be
rewritten, an error will be thrown during tag generation when the empty
ESC collection namespace is encountered. If a runtime comparison is
forced instead, the predicate will fail to match on any documents since
the documents don't contain encrypted data.

In the very unlikely event that a pipeline involved an encrypted
collection, but we could not find an entry in `encryptionInformation`,
an error will be thrown during tag generation when the empty ESC
collection namespace is encountered. If a runtime comparison is forced
instead, the predicate will still match the expected documents, since
the runtime comparison only needs the bindata payload to perform the
predicate evaluation and does not require the encryption schema.

Under no circumstances is any sensitive plaintext data exposed.