Add desugaring stages to the plugin (#33130)
Desugaring stages have no concrete implementation and simply expand to
one or more other stages. Search stages
typically expand into at least an underlying remote stage and an
idLookup stage and this is generally useful to be able
to control independently of concrete stages. This could also be used to
replace the manual desugaring expansions that
happen directly in agg pipeline execution today.