SERVER-101198 Remove dead code from PCRE wrapper (#33210)
Previously, the pcre::MatchDataImpl wrapper class had a size_t instance
variable `_highestCaptureIndex`, which was populated with the value of a
signed integer, which was potentially negative. This could have led to a
wrong value being stored in that variable in case the call to the PCRE2
regex match function returned an error.
However, the `_highestCaptureIndex` variable was only populated and
never read back, so it is safe to remove it.

The PR also adds some code comment to the wrapper function that calls
`pcre2_match()´, so that it can be better understood what the different
return values mean.
Finally, it adds some unit tests for a few common cases.