SERVER-99915 FLE2 findAndModify should also return WCE from transaction API result (#32689)
For cases where a FLE2 findAndModify command encounters both an error
and a write concern error, it now uses the `ErrorWithWriteConcernError`
status code (added in SERVER-100435) so that it can report both its
internal WCE, and the actual error status in the command reply.

The `ErrorWithWriteConcernError` status code is for internal use only,
and is not meant to be sent in command replies to users. SERVER-100435
added code in `appendCommandStatusNoThrow()` that will catch this and
unpack the actual error code & WCE to be included in the reply, but that
only worked for mongos. mongod uses a different code path for building
its command replies (`augmentReplyWithStatus()` in
`reply_builder_interface.cpp`, which is supposed to be a duplicate of
`appendCommandStatusNoThrow` for mongod), but that did not get updated
in SERVER-100435. This patch adds the same changes that SERVER-100435
added to `appendCommandStatusNoThrow()`, to `augmentReplyWithStatus()`.