SERVER-101804 fix plan cache issue when routing info is stale (#33160)
In index_filter_commands.js, plan_cache_clear.js and
plan_cache_list_plans.js, we run the lookup query first, then run a
explain for that query to get plan cache key, and use the key to
retrieve from plan cache, this is usually fine because we always
generate same plan for that query.

But sometime this assumption is wrong because the node may have stale
routing info, foreign coll might be treated as non-local incorrectly,
the SBE is not used for the lookup non-local foreign coll, but after
routing info refreshed SBE might be used for the second explain query.
(The lookup for the foreign coll fails first time due to stale info, but
will be retried, so we can get the result successfully)