SERVER-100922: Track AWS SDK response codes (Addresses auto-revert) (#33255)
Summary:

- Introduces the same changes from #32880 , but addresses #33203 and
additional change
- #33203 was merged as a result of lambda_local.js test hanging. It was
found that this was due to mongostream hanging during shutdown while
taking snapshot of metrics. This is because, the new CounterVec was
stuck in an infinite loop while taking snapshots of all its counters
without ever advancing it's iterator
- Additional change: I found that we originally wouldn't record error
response codes or responses occurring within the sink operator. The
added tests failed to confirm this as they failed to take into account
that it must build AWS lambda errors for the SDK to return them as an
error.