Add agg pipeline context as bson when creating agg stages (#33282)
Agg pipeline context is serialized as bson by the plugin host and passed
to plugins when parsing concrete agg stages.
This includes the contents of the namespace string, collection UUID, and
a bit indicating if we are processing the pipeline
in a routing role (~mongos) or data role (~mongod). Correct a number of
TODOs related to passing and reading this
context before sending remote requests.

At the moment we have some trivial stages that don't use it but we may
want to parse context pre-emptively if it
seems like most will use it.

Make some smaller changes:
* Fix some API names in the header to correct some of the failing lint
checks.
* Make the rust `Error` case accept `Into<String>` so we get some
overload-ish behavior that will improve ergonomics.
* Remove unused imports in the rust code.