SERVER-101882: include AFs in #10gen-mongo-code-lockdown (#33224)
We're working to include AFs in the quota system.
Adding a new category of things makes the prioritisation harder, so
we've decided we'd like to collapse the reporting into simply hot vs
cold issues, and simplify the quotas accordingly.

This updates the reporting to take that into account.

It also offloads the logic for identifying issues to saved filters in
Jira, which also back the dashboards that allow people to drill in which
should simplify the maintenance of which issues are / aren't included,
or which are hot and which are cold.