SERVER-101956: Minor cleanup related to WiredTigerServerStatusSection. (#33334)
Currently, all 3 implementations of `StorageEngine::Factory` use
`kWiredTigerEngineName` to specify the name of the serverStatus section
for the storage engine. However, this is a bit clunky since
`kWiredTigerEngineName` is also used to specify the config file section
name for `WiredTigerKVEngine`, similar to how ‘inMemory’ is used for
`InMemoryFactory` and `queryable_wt` for `QueryableWtFactory`. This
caused a bit of confusion for me. Fixing it by using a separate string
literal for `WiredTigerServerStatusSection` to decouple it from the
string literals used for the config file.