SERVER-101735: WiredTigerKVEngine supports creating WiredTiger instances with different configurations. (#33152)
This is basically a code refactoring pr with no changes to
functionality.

Changes in this pr:
- Supply configuration params to `WiredTigerKVEngine` via the new
`WiredTigerConfig` struct.
- Supply configuration params to `WiredTigerRecordStore` via the new
`WiredTigerTableConfig` struct.
- Rename `StorageGlobalParams.ephemeral` to
`StorageGlobalParams.inMemory`. `ephemeral` keyword (and hence
`StorageEngine::isEphemeral()`) will now indicate one of the following -
either the storage engine is in-memory or it is non-durable i.e. it can
lose its data on disk upon a restart. We will add support for the
"non-durable" functionality in future PRs.
- Delete unused
`WiredTigerGlobalOptions::useCollectionPrefixCompression`.

Note that we are still not creating any more WiredTiger instances than
we do today. This pr is simply taking us one step closer to being able
to do that.