SERVER-101321 Fix potential uint32_t overflow (#33231)
Prevent a potential uint32_t overflow in `BSONObj::woCompare()` when the
number of fields to compare exceeds 31.
The uint32_t value was used to create a bitmask for calling
`Ordering::descending(mask)`.

It turns out that an `Ordering` can only have up to 32 values (asc /
desc), but it is ok to query the order of extraneous offsets via the
already existing `Ordering::get(int offset)` function. This function
returns `1` for all extraneous offsets.
In order to prevent the potential overflow in `BSONObj::woCompare()` and
in order to unify the behavior, this PR makes `BSONObj::woCompare()` now
call the safe `Ordering::get(int offset)` function instead of
`Ordering::descending(uint32_t mask)`.

---------

Co-authored-by: Billy Donahue <BillyDonahue@users.noreply.github.com>