SERVER-101735: Consolidate all wiredtiger config options and server parameters in wiredtiger_global_options.idl. (#33351)
Changes in this pr:
- As done in `mongod_options_storage.idl`, consolidate all wiredtiger
parameters and config options in wiredtiger_global_options.idl. The
current separation has already caused a great deal of back-and-forth for
me.
- As requested in https://github.com/10gen/mongo/pull/33152, replace
`gWiredTigerSessionMax` with `wiredTigerGlobalOptions.sessionMax`,
`gWiredTigerEvictionDirtyTargetGB` with
`wiredTigerGlobalOptions.evictionDirtyTargetGB`, and
`gWiredTigerEvictionDirtyMaxGB` with
`wiredTigerGlobalOptions.evictionDirtyTriggerGB`. We could do this for
more parameters, but we should do it on a need basis.