SERVER-101591 [SBE] Release storage engine resources when saveState() or restoreState() throws (#32975)
This PR adds logic to ensure that storage resources (CollectionPtrs and
cursors) are eagerly released when saveState() or restoreState() throw
an exception. This is needed to satisfy invariants in
~CollectionsInUse() in the event that the snapshot is abandoned before
the SBE tree can be disposed of (see BF-36692 for details).

This PR also fixes CollectionRef so that a CollectionPtr is not held if
an exception is thrown from CollectionRef::acquireCollection() or
CollectionRef::restoreCollection().