SERVER-102031 numBucketsOpenedDueToMetadata always increases (#33356)
- Removed the `bucketOpenedDueToMetadata` arg/setting completely from
the `stageInsertBatchIntoEligibleBucket` path and moved it to
`getEligibleBucket`/`stageInsertBatch` path.
- Passing `bucketOpenedDueToMetadata` as a ref-bool that we set in
`getEligibleBucket` and its helper function so we can appropriately set
this in `stageInsertBatch`.
- Added assertion in happy path testing to check that
`bucketOpenedDueToMetadata` is properly set
- Moved the call to `activeBatch` into `stageInsertBatch` rather than
`stageInsertBatchIntoEligibleBucket` so we can directly set
`openedDueToMetadata` within `stageInsertBatch`

Tracing through the old code path, we set `batch->openedDueToMetadata =
true`; iff we have (for any of the three cases; archival-based
reopening, query-based reopening, or ranging through the set of open
buckets by bucket key):
1.
[`isNewlyOpenedBucket`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L1829)
is true -- simply if we have a bucket with no measurements
2.
[`openedDueToMetadata`](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L1831)
is true -- this is set to true when we know that we don't have a newly
opened bucket, but we then proceed with rollover (which in the old path
allocates a new bucket)
3. The bucket was successfully loaded into the catalog.
This two conditionals in combination mean: We have to have a bucket that
is allocated, where there were no candidate buckets that were able to be
loaded into the catalog successfully from looking at the open set of
buckets for the bucket key, from archive-based reopening, or query based
reopening.

For Yuhong -- in the case you were talking about where we have 5
separate measurements in 5 separate batches, we don't increment every
time we have a `RolloverReason::kNone` because we will have a non-empty
bucket to insert into, so we won't ever set `batch->openedDueToMetadata
`(condition 1).

Second note for Yuhong, we use `isNewlyOpenedBucket` for schema check
[here](https://github.com/10gen/mongo/blob/master/src/mongo/db/timeseries/bucket_catalog/bucket_catalog_internal.cpp#L1834),
so I believe we still want to keep this until/when we proceed with
(SERVER-100028)[https://jira.mongodb.org/browse/SERVER-100028].

---------

Co-authored-by: Matt Kneiser <matt.kneiser@mongodb.com>