SERVER-102055 Remove TimeseriesBucketNamespace wrapper class (#33362)
TimeseriesBucketNamespace was created to satisfy the requirements of
`timeseries::isTimeseriesViewRequest.`

The proposal is instead to adapt the requirements of
`timeseries::isTimeseriesViewRequest` for `NamespaceInfoEntry`. In fact
this class already have all the information needed for
`isTimeseriesViewRequest`. It has a namespace and also the
`isTimeseriesNamespace` flag.

The results is simpler code on the `bulk_write.cpp`.