SERVER-102117 Remove WiredTigerStatsTest::EmptySession (#33426)
This test asserts that the read and write statistics are empty. However,
when WiredTiger opens a btree that was checkpointed it can warm up the
cache by reading some of the btree content. This makes the test
assertion flakey. There's no way to turn off the cache pre-loading, so
we should remove the test.