SERVER-101553 Expose commitTimestamp for prepared transactions in change stream events (#33243)
Expose a `commitTimestamp` field in CRUD change stream events that are
part of a prepared transaction, when the change stream is opened with
`showExpandedEvents: true`. (note: all cross-shard transactions are
implicitly prepared)
If set, the `commitTimestamp` field in the event resembles the value of
the `commitTimestamp` field from the `commitTransaction` oplog entry.